fix(init): harden Wrangler TOML config updates
vinext init could treat valid but unsupported Wrangler TOML owned-field shapes as missing fields. That allowed dotted keys or inline KV namespace arrays to be followed by duplicate generated config instead of failing before mutation.

The config update path now builds one existing-Wrangler update plan for JSON and TOML, and init uses that plan for both validation and setup. TOML edits stay source-preserving for inline tables, normal tables, and KV array tables, while unsupported owned shapes are rejected before any project files or install commands are touched.
