+++ packages/vinext/src/init-cloudflare.ts
@@ -5,6 +5,7 @@ import MagicString from "magic-string";
 import type { ESTree } from "vite";
 import type { CloudflareInitOptions } from "./init-platform.js";
 import { detectProject } from "./utils/project.js";
+import { createExistingWranglerConfigUpdatePlan } from "./wrangler-config.js";
 
 const require = createRequire(import.meta.url);
 
@@ -21,17 +22,6 @@ const DEFAULT_CLOUDFLARE_INIT_OPTIONS: CloudflareInitOptions = {
   dataCache: "kv",
   cdnCache: "workers-cache",
   imageOptimization: "cloudflare-images",
-};
-
-type WranglerConfigFormat = "json" | "toml";
-
-type ExistingWranglerConfig = {
-  path: string;
-  format: WranglerConfigFormat;
-};
-
-type ExistingWranglerConfigFile = ExistingWranglerConfig & {
-  code: string;
 };
 
 export type CloudflarePlatformSetupContext = {
@@ -54,16 +44,8 @@ export function validateCloudflarePlatformSetup(
   cloudflare: CloudflareInitOptions,
 ): void {
   const projectInfo = detectProject(context.root);
-  const wranglerConfig = readExistingWranglerConfig(context.root);
-  let imagesBinding = "IMAGES";
-  if (wranglerConfig) {
-    const updatedWranglerCode = updateExistingWranglerConfigForCloudflare(
-      wranglerConfig,
-      wranglerConfig.code,
-      cloudflare,
-    );
-    imagesBinding = getExistingWranglerImagesBinding(wranglerConfig, updatedWranglerCode);
-  }
+  const wranglerConfigPlan = createExistingWranglerConfigUpdatePlan(context.root, cloudflare);
+  const imagesBinding = wranglerConfigPlan?.imagesBinding ?? "IMAGES";
 
   if (context.existingViteConfigPath) {
     updateViteConfigForCloudflare(
@@ -85,10 +67,8 @@ export function setupCloudflarePlatform(
   cloudflare: CloudflareInitOptions,
 ): CloudflarePlatformSetupResult {
   const projectInfo = detectProject(context.root);
-  const wranglerConfig = readExistingWranglerConfig(context.root);
-  const imagesBinding = wranglerConfig
-    ? getExistingWranglerImagesBinding(wranglerConfig, wranglerConfig.code)
-    : "IMAGES";
+  const wranglerConfigPlan = createExistingWranglerConfigUpdatePlan(context.root, cloudflare);
+  const imagesBinding = wranglerConfigPlan?.imagesBinding ?? "IMAGES";
 
   let generatedViteConfig = false;
   let skippedViteConfig = false;
@@ -120,7 +100,9 @@ export function setupCloudflarePlatform(
   }
 
   const generatedPlatformFiles: string[] = [];
-  if (!wranglerConfig) {
+  let finalWranglerFileName = "wrangler.jsonc";
+  let needsKvNamespaceId = cloudflare.dataCache === "kv";
+  if (!wranglerConfigPlan) {
     fs.writeFileSync(
       path.join(context.root, "wrangler.jsonc"),
       generateWranglerConfig(projectInfo, cloudflare, context.today),
@@ -128,28 +110,13 @@ export function setupCloudflarePlatform(
     );
     generatedPlatformFiles.push("wrangler.jsonc");
   } else {
-    const updatedConfig = updateExistingWranglerConfigForCloudflare(
-      wranglerConfig,
-      wranglerConfig.code,
-      cloudflare,
-    );
-    if (updatedConfig !== wranglerConfig.code) {
-      fs.writeFileSync(wranglerConfig.path, updatedConfig, "utf-8");
-      generatedPlatformFiles.push(path.basename(wranglerConfig.path));
+    if (wranglerConfigPlan.changed) {
+      fs.writeFileSync(wranglerConfigPlan.path, wranglerConfigPlan.code, "utf-8");
+      generatedPlatformFiles.push(wranglerConfigPlan.fileName);
     }
+    finalWranglerFileName = wranglerConfigPlan.fileName;
+    needsKvNamespaceId = wranglerConfigPlan.needsKvNamespaceId;
   }
-
-  const finalWranglerConfig = wranglerConfig ?? {
-    path: path.join(context.root, "wrangler.jsonc"),
-    format: "json" as const,
-  };
-  const finalWranglerFileName = path.basename(finalWranglerConfig.path);
-  const finalWranglerCode = fs.readFileSync(finalWranglerConfig.path, "utf-8");
-  const needsKvNamespaceId = needsWranglerKvNamespaceId(
-    finalWranglerConfig,
-    finalWranglerCode,
-    cloudflare,
-  );
 
   return {
     generatedViteConfig,
@@ -165,72 +132,6 @@ export function setupCloudflarePlatform(
         ]
       : [],
   };
-}
-
-function findExistingWranglerConfig(root: string): ExistingWranglerConfig | undefined {
-  for (const fileName of ["wrangler.jsonc", "wrangler.json", "wrangler.toml"]) {
-    const candidate = path.join(root, fileName);
-    if (!fs.existsSync(candidate)) continue;
-    return {
-      path: candidate,
-      format: fileName.endsWith(".toml") ? "toml" : "json",
-    };
-  }
-  return undefined;
-}
-
-function readExistingWranglerConfig(root: string): ExistingWranglerConfigFile | undefined {
-  const config = findExistingWranglerConfig(root);
-  if (!config) return undefined;
-  return { ...config, code: fs.readFileSync(config.path, "utf-8") };
-}
-
-function updateExistingWranglerConfigForCloudflare(
-  config: ExistingWranglerConfig,
-  code: string,
-  options: CloudflareInitOptions,
-): string {
-  if (config.format === "json") {
-    return updateWranglerConfigForCloudflare(code, options);
-  }
-  try {
-    return updateWranglerTomlConfigForCloudflare(code, options);
-  } catch (cause) {
-    throw new Error("Could not update the existing Wrangler TOML config.", { cause });
-  }
-}
-
-function getExistingWranglerImagesBinding(config: ExistingWranglerConfig, code: string): string {
-  if (config.format === "json") {
-    return getWranglerImagesBinding(code);
-  }
-  try {
-    return getWranglerTomlImagesBinding(code);
-  } catch (cause) {
-    throw new Error("Could not update the existing Wrangler TOML config.", { cause });
-  }
-}
-
-function needsWranglerKvNamespaceId(
-  config: ExistingWranglerConfig,
-  code: string,
-  options: CloudflareInitOptions,
-): boolean {
-  if (options.dataCache !== "kv") return false;
-  if (config.format === "toml") return tomlKvNamespaceNeedsId(code);
-
-  const finalWranglerConfig = JSON.parse(stripJsonComments(code)) as {
-    kv_namespaces?: Array<{ binding?: unknown; id?: unknown }>;
-  };
-  const kvBinding = finalWranglerConfig.kv_namespaces?.find(
-    (namespace) => namespace.binding === "VINEXT_KV_CACHE",
-  );
-  return (
-    !kvBinding ||
-    typeof kvBinding.id !== "string" ||
-    kvBinding.id.length === 0 ||
-    kvBinding.id === "<your-kv-namespace-id>"
-  );
 }
 
 // Cloudflare deployment scaffolding belongs to `vinext init`.
@@ -275,639 +176,6 @@ export function generateWranglerConfig(
   }
 
   return JSON.stringify(config, null, 2) + "\n";
-}
-
-function stripJsonComments(code: string): string {
-  let output = "";
-  let inString = false;
-  let escaped = false;
-  for (let index = 0; index < code.length; index++) {
-    const char = code[index];
-    const next = code[index + 1];
-    if (inString) {
-      output += char;
-      if (escaped) escaped = false;
-      else if (char === "\\") escaped = true;
-      else if (char === '"') inString = false;
-      continue;
-    }
-    if (char === '"') {
-      inString = true;
-      output += char;
-      continue;
-    }
-    if (char === "/" && next === "/") {
-      while (index < code.length && code[index] !== "\n") index++;
-      output += "\n";
-      continue;
-    }
-    if (char === "/" && next === "*") {
-      index += 2;
-      while (index < code.length && !(code[index] === "*" && code[index + 1] === "/")) {
-        output += code[index] === "\n" ? "\n" : " ";
-        index++;
-      }
-      index++;
-      continue;
-    }
-    output += char;
-  }
-  return output.replace(/,\s*([}\]])/g, "$1");
-}
-
-function findTopLevelJsonProperty(
-  code: string,
-  name: string,
-): { valueStart: number; valueEnd: number } | null {
-  let depth = 0;
-  let inString = false;
-  let escaped = false;
-  let lineComment = false;
-  let blockComment = false;
-  for (let index = 0; index < code.length; index++) {
-    const char = code[index];
-    const next = code[index + 1];
-    if (lineComment) {
-      if (char === "\n") lineComment = false;
-      continue;
-    }
-    if (blockComment) {
-      if (char === "*" && next === "/") {
-        blockComment = false;
-        index++;
-      }
-      continue;
-    }
-    if (inString) {
-      if (escaped) escaped = false;
-      else if (char === "\\") escaped = true;
-      else if (char === '"') inString = false;
-      continue;
-    }
-    if (char === "/" && next === "/") {
-      lineComment = true;
-      index++;
-      continue;
-    }
-    if (char === "/" && next === "*") {
-      blockComment = true;
-      index++;
-      continue;
-    }
-    if (char === '"') {
-      inString = true;
-      let value = "";
-      index++;
-      for (; index < code.length; index++) {
-        const stringChar = code[index];
-        if (stringChar === "\\") {
-          value += stringChar + (code[++index] ?? "");
-        } else if (stringChar === '"') {
-          inString = false;
-          break;
-        } else value += stringChar;
-      }
-      if (depth !== 1 || value !== name) continue;
-      let cursor = index + 1;
-      while (/\s/.test(code[cursor] ?? "")) cursor++;
-      if (code[cursor] !== ":") continue;
-      cursor++;
-      while (/\s/.test(code[cursor] ?? "")) cursor++;
-      const valueStart = cursor;
-      let valueDepth = 0;
-      let valueString = false;
-      let valueEscaped = false;
-      let valueLineComment = false;
-      let valueBlockComment = false;
-      for (; cursor < code.length; cursor++) {
-        const valueChar = code[cursor];
-        const valueNext = code[cursor + 1];
-        if (valueLineComment) {
-          if (valueChar === "\n") valueLineComment = false;
-          continue;
-        }
-        if (valueBlockComment) {
-          if (valueChar === "*" && valueNext === "/") {
-            valueBlockComment = false;
-            cursor++;
-          }
-          continue;
-        }
-        if (valueString) {
-          if (valueEscaped) valueEscaped = false;
-          else if (valueChar === "\\") valueEscaped = true;
-          else if (valueChar === '"') valueString = false;
-          continue;
-        }
-        if (valueChar === "/" && valueNext === "/") {
-          valueLineComment = true;
-          cursor++;
-        } else if (valueChar === "/" && valueNext === "*") {
-          valueBlockComment = true;
-          cursor++;
-        } else if (valueChar === '"') valueString = true;
-        else if (valueChar === "{" || valueChar === "[") valueDepth++;
-        else if (valueChar === "}" || valueChar === "]") {
-          if (valueDepth === 0) return { valueStart, valueEnd: cursor };
-          valueDepth--;
-          if (valueDepth === 0) return { valueStart, valueEnd: cursor + 1 };
-        } else if (valueChar === "," && valueDepth === 0) {
-          return { valueStart, valueEnd: cursor };
-        }
-      }
-      return { valueStart, valueEnd: cursor };
-    }
-    if (char === "{") depth++;
-    else if (char === "}") depth--;
-  }
-  return null;
-}
-
-function appendTopLevelJsonProperty(code: string, property: string): string {
-  const closing = code.lastIndexOf("}");
-  if (closing < 0) throw new Error("Could not find the root object in Wrangler config.");
-  const before = code.slice(0, closing);
-  const structuralBefore = stripJsonComments(before);
-  const needsComma = !/,\s*$/.test(structuralBefore) && !/{\s*$/.test(structuralBefore);
-  return `${before}${needsComma ? "," : ""}\n${property}\n${code.slice(closing)}`;
-}
-
-export function updateWranglerConfigForCloudflare(
-  code: string,
-  options: CloudflareInitOptions,
-): string {
-  try {
-    JSON.parse(stripJsonComments(code));
-  } catch (cause) {
-    throw new Error("Could not parse the existing Wrangler JSON/JSONC config.", { cause });
-  }
-  let output = code;
-  if (options.cdnCache === "workers-cache") {
-    const cacheProperty = findTopLevelJsonProperty(output, "cache");
-    if (!cacheProperty) {
-      output = appendTopLevelJsonProperty(output, '  "cache": { "enabled": true }');
-    } else {
-      const cache = JSON.parse(
-        stripJsonComments(output.slice(cacheProperty.valueStart, cacheProperty.valueEnd)),
-      ) as Record<string, unknown> | null;
-      if (!cache || cache.enabled !== true) {
-        const updatedCache = JSON.stringify({ ...cache, enabled: true });
-        output = `${output.slice(0, cacheProperty.valueStart)}${updatedCache}${output.slice(cacheProperty.valueEnd)}`;
-      }
-    }
-  }
-  if (options.imageOptimization === "cloudflare-images") {
-    const imagesProperty = findTopLevelJsonProperty(output, "images");
-    if (!imagesProperty) {
-      output = appendTopLevelJsonProperty(output, '  "images": { "binding": "IMAGES" }');
-    } else {
-      const images = JSON.parse(
-        stripJsonComments(output.slice(imagesProperty.valueStart, imagesProperty.valueEnd)),
-      ) as { binding?: unknown } | null;
-      if (!images || typeof images.binding !== "string" || images.binding.length === 0) {
-        output = `${output.slice(0, imagesProperty.valueStart)}{ "binding": "IMAGES" }${output.slice(imagesProperty.valueEnd)}`;
-      }
-    }
-  }
-  if (options.dataCache === "kv") {
-    const kvProperty = findTopLevelJsonProperty(output, "kv_namespaces");
-    if (!kvProperty) {
-      output = appendTopLevelJsonProperty(
-        output,
-        '  "kv_namespaces": [{ "binding": "VINEXT_KV_CACHE", "id": "<your-kv-namespace-id>" }]',
-      );
-    } else {
-      const rawValue = output.slice(kvProperty.valueStart, kvProperty.valueEnd);
-      const namespaces = JSON.parse(stripJsonComments(rawValue)) as Array<{ binding?: string }>;
-      if (!namespaces.some((namespace) => namespace.binding === "VINEXT_KV_CACHE")) {
-        const closing = kvProperty.valueEnd - 1;
-        const content = output.slice(kvProperty.valueStart + 1, closing);
-        const separator = content.trim() ? `${/,\s*$/.test(content) ? "" : ","}\n    ` : "";
-        output = `${output.slice(0, closing)}${separator}{ "binding": "VINEXT_KV_CACHE", "id": "<your-kv-namespace-id>" }${output.slice(closing)}`;
-      }
-    }
-  }
-  return output;
-}
-
-export function getWranglerImagesBinding(code: string): string {
-  const property = findTopLevelJsonProperty(code, "images");
-  if (!property) return "IMAGES";
-  const images = JSON.parse(
-    stripJsonComments(code.slice(property.valueStart, property.valueEnd)),
-  ) as { binding?: unknown } | null;
-  return images && typeof images.binding === "string" && images.binding.length > 0
-    ? images.binding
-    : "IMAGES";
-}
-
-type TomlAssignment = {
-  valueStart: number;
-  valueEnd: number;
-  value: string;
-};
-
-type TomlSection = {
-  name: string;
-  isArray: boolean;
-  bodyStart: number;
-  bodyEnd: number;
-};
-
-// This is a targeted Wrangler TOML updater, not a general TOML serializer.
-// It only edits the top-level fields vinext init owns and throws when those
-// fields have shapes that cannot be patched without rewriting user config.
-function stripTomlLineComment(line: string): string {
-  let quote: "'" | '"' | undefined;
-  let escaped = false;
-  for (let index = 0; index < line.length; index++) {
-    const char = line[index];
-    if (quote === '"') {
-      if (escaped) escaped = false;
-      else if (char === "\\") escaped = true;
-      else if (char === '"') quote = undefined;
-      continue;
-    }
-    if (quote === "'") {
-      if (char === "'") quote = undefined;
-      continue;
-    }
-    if (char === '"' || char === "'") {
-      quote = char;
-      continue;
-    }
-    if (char === "#") return line.slice(0, index);
-  }
-  return line;
-}
-
-function parseTomlHeader(line: string): { name: string; isArray: boolean } | undefined {
-  const trimmed = stripTomlLineComment(line).trim();
-  const match = trimmed.match(/^(\[\[?)\s*([^[\]]+?)\s*(\]\]?)$/);
-  if (!match) return undefined;
-  const isArray = match[1] === "[[";
-  if (isArray !== (match[3] === "]]")) return undefined;
-  return { name: match[2].trim(), isArray };
-}
-
-function forEachTomlLine(
-  code: string,
-  callback: (line: string, lineStart: number, lineEnd: number) => void,
-): void {
-  let lineStart = 0;
-  while (lineStart <= code.length) {
-    const newline = code.indexOf("\n", lineStart);
-    const lineEndWithoutNewline = newline === -1 ? code.length : newline;
-    const lineEnd = newline === -1 ? code.length : newline + 1;
-    callback(code.slice(lineStart, lineEndWithoutNewline), lineStart, lineEnd);
-    if (newline === -1) break;
-    lineStart = newline + 1;
-  }
-}
-
-function findFirstTomlSectionStart(code: string): number | undefined {
-  let firstSectionStart: number | undefined;
-  forEachTomlLine(code, (line, lineStart) => {
-    if (firstSectionStart !== undefined) return;
-    if (parseTomlHeader(line)) firstSectionStart = lineStart;
-  });
-  return firstSectionStart;
-}
-
-function findTomlSections(code: string, name: string, isArray: boolean): TomlSection[] {
-  const sections: TomlSection[] = [];
-  let current: TomlSection | undefined;
-  forEachTomlLine(code, (line, lineStart, lineEnd) => {
-    const header = parseTomlHeader(line);
-    if (!header) return;
-    if (current) {
-      current.bodyEnd = lineStart;
-      sections.push(current);
-    }
-    current = {
-      name: header.name,
-      isArray: header.isArray,
-      bodyStart: lineEnd,
-      bodyEnd: code.length,
-    };
-  });
-  if (current) sections.push(current);
-  return sections.filter((section) => section.name === name && section.isArray === isArray);
-}
-
-function findTomlAssignmentInRange(
-  code: string,
-  name: string,
-  start: number,
-  end: number,
-): TomlAssignment | undefined {
-  let match: TomlAssignment | undefined;
-  forEachTomlLine(code.slice(start, end), (line, relativeLineStart) => {
-    if (match) return;
-    const uncommented = stripTomlLineComment(line);
-    const keyMatch = uncommented.match(/^\s*([A-Za-z0-9_-]+)\s*=/);
-    if (!keyMatch || keyMatch[1] !== name) return;
-
-    let relativeValueStart = keyMatch[0].length;
-    while (/\s/.test(uncommented[relativeValueStart] ?? "")) relativeValueStart++;
-    let relativeValueEnd = uncommented.length;
-    while (relativeValueEnd > relativeValueStart && /\s/.test(uncommented[relativeValueEnd - 1])) {
-      relativeValueEnd--;
-    }
-
-    match = {
-      valueStart: start + relativeLineStart + relativeValueStart,
-      valueEnd: start + relativeLineStart + relativeValueEnd,
-      value: code.slice(
-        start + relativeLineStart + relativeValueStart,
-        start + relativeLineStart + relativeValueEnd,
-      ),
-    };
-  });
-  return match;
-}
-
-function findTopLevelTomlAssignment(code: string, name: string): TomlAssignment | undefined {
-  return findTomlAssignmentInRange(code, name, 0, findFirstTomlSectionStart(code) ?? code.length);
-}
-
-function parseTomlString(value: string): string | undefined {
-  const trimmed = value.trim();
-  if (trimmed.startsWith('"') && trimmed.endsWith('"')) {
-    try {
-      const parsed = JSON.parse(trimmed);
-      return typeof parsed === "string" ? parsed : undefined;
-    } catch {
-      return undefined;
-    }
-  }
-  if (trimmed.startsWith("'") && trimmed.endsWith("'")) return trimmed.slice(1, -1);
-  return undefined;
-}
-
-function parseTomlBoolean(value: string): boolean | undefined {
-  const trimmed = value.trim();
-  if (trimmed === "true") return true;
-  if (trimmed === "false") return false;
-  return undefined;
-}
-
-function isInlineTomlObject(value: string): boolean {
-  const trimmed = value.trim();
-  return trimmed.startsWith("{") && trimmed.endsWith("}");
-}
-
-function findInlineTomlProperty(
-  value: string,
-  name: string,
-): { valueStart: number; valueEnd: number; value: string } | undefined {
-  const open = value.indexOf("{");
-  const close = value.lastIndexOf("}");
-  if (open < 0 || close < open) return undefined;
-
-  let cursor = open + 1;
-  while (cursor < close) {
-    while (/[\s,]/.test(value[cursor] ?? "") && cursor < close) cursor++;
-    if (cursor >= close) break;
-
-    const keyStart = cursor;
-    while (/[A-Za-z0-9_-]/.test(value[cursor] ?? "")) cursor++;
-    const key = value.slice(keyStart, cursor);
-    while (/\s/.test(value[cursor] ?? "") && cursor < close) cursor++;
-    if (value[cursor] !== "=") return undefined;
-    cursor++;
-    while (/\s/.test(value[cursor] ?? "") && cursor < close) cursor++;
-
-    const valueStart = cursor;
-    let quote: "'" | '"' | undefined;
-    let escaped = false;
-    let nestedDepth = 0;
-    for (; cursor < close; cursor++) {
-      const char = value[cursor];
-      if (quote === '"') {
-        if (escaped) escaped = false;
-        else if (char === "\\") escaped = true;
-        else if (char === '"') quote = undefined;
-        continue;
-      }
-      if (quote === "'") {
-        if (char === "'") quote = undefined;
-        continue;
-      }
-      if (char === '"' || char === "'") {
-        quote = char;
-      } else if (char === "{" || char === "[") {
-        nestedDepth++;
-      } else if (char === "}" || char === "]") {
-        if (nestedDepth > 0) nestedDepth--;
-      } else if (char === "," && nestedDepth === 0) {
-        break;
-      }
-    }
-    let valueEnd = cursor;
-    while (valueEnd > valueStart && /\s/.test(value[valueEnd - 1])) valueEnd--;
-    if (key === name) {
-      return {
-        valueStart,
-        valueEnd,
-        value: value.slice(valueStart, valueEnd),
-      };
-    }
-    if (value[cursor] === ",") cursor++;
-  }
-  return undefined;
-}
-
-function addInlineTomlProperty(value: string, property: string): string {
-  const close = value.lastIndexOf("}");
-  if (close < 0) throw new Error("Expected an inline TOML object.");
-  const open = value.indexOf("{");
-  const content = value.slice(open + 1, close);
-  const separator =
-    content.trim().length > 0 ? `${content.trimEnd().endsWith(",") ? "" : ","} ` : " ";
-  return `${value.slice(0, close)}${separator}${property} ${value.slice(close)}`;
-}
-
-function replaceTomlAssignmentValue(
-  code: string,
-  assignment: TomlAssignment,
-  value: string,
-): string {
-  return `${code.slice(0, assignment.valueStart)}${value}${code.slice(assignment.valueEnd)}`;
-}
-
-function appendTomlTopLevelAssignment(code: string, assignment: string): string {
-  const firstSectionStart = findFirstTomlSectionStart(code);
-  if (firstSectionStart === undefined) {
-    const separator = code.length === 0 || code.endsWith("\n") ? "" : "\n";
-    return `${code}${separator}${assignment}\n`;
-  }
-
-  let before = code.slice(0, firstSectionStart);
-  const after = code.slice(firstSectionStart);
-  if (before.length > 0 && !before.endsWith("\n")) before += "\n";
-  const separatorBefore = before.trim().length > 0 && !before.endsWith("\n\n") ? "\n" : "";
-  const separatorAfter = after.length > 0 ? "\n" : "";
-  return `${before}${separatorBefore}${assignment}\n${separatorAfter}${after}`;
-}
-
-function appendTomlArrayTable(code: string, table: string): string {
-  const separator =
-    code.length === 0 ? "" : code.endsWith("\n\n") ? "" : code.endsWith("\n") ? "\n" : "\n\n";
-  return `${code}${separator}${table}\n`;
-}
-
-function setInlineTomlPropertyValue(
-  code: string,
-  assignment: TomlAssignment,
-  name: string,
-  value: string,
-): string {
-  const property = findInlineTomlProperty(assignment.value, name);
-  if (property) {
-    return `${code.slice(0, assignment.valueStart + property.valueStart)}${value}${code.slice(
-      assignment.valueStart + property.valueEnd,
-    )}`;
-  }
-  return replaceTomlAssignmentValue(
-    code,
-    assignment,
-    addInlineTomlProperty(assignment.value, `${name} = ${value}`),
-  );
-}
-
-function setTomlTableAssignmentValue(
-  code: string,
-  section: TomlSection,
-  name: string,
-  value: string,
-): string {
-  const assignment = findTomlAssignmentInRange(code, name, section.bodyStart, section.bodyEnd);
-  if (assignment) return replaceTomlAssignmentValue(code, assignment, value);
-
-  const insertion = `${name} = ${value}\n`;
-  return `${code.slice(0, section.bodyEnd)}${code[section.bodyEnd - 1] === "\n" ? "" : "\n"}${insertion}${code.slice(section.bodyEnd)}`;
-}
-
-function ensureTomlCacheEnabled(code: string): string {
-  const inlineCache = findTopLevelTomlAssignment(code, "cache");
-  const tableCache = findTomlSections(code, "cache", false)[0];
-  if (inlineCache && tableCache) throw new Error("Wrangler TOML defines cache twice.");
-
-  if (inlineCache) {
-    if (!isInlineTomlObject(inlineCache.value)) {
-      throw new Error("Expected top-level cache to be a TOML object.");
-    }
-    const enabled = findInlineTomlProperty(inlineCache.value, "enabled");
-    if (enabled && parseTomlBoolean(enabled.value) === true) return code;
-    return setInlineTomlPropertyValue(code, inlineCache, "enabled", "true");
-  }
-
-  if (tableCache) {
-    const enabled = findTomlAssignmentInRange(
-      code,
-      "enabled",
-      tableCache.bodyStart,
-      tableCache.bodyEnd,
-    );
-    if (enabled && parseTomlBoolean(enabled.value) === true) return code;
-    return setTomlTableAssignmentValue(code, tableCache, "enabled", "true");
-  }
-
-  return appendTomlTopLevelAssignment(code, "cache = { enabled = true }");
-}
-
-function getWranglerTomlImagesBinding(code: string): string {
-  const inlineImages = findTopLevelTomlAssignment(code, "images");
-  if (inlineImages && isInlineTomlObject(inlineImages.value)) {
-    const binding = findInlineTomlProperty(inlineImages.value, "binding");
-    const value = binding ? parseTomlString(binding.value) : undefined;
-    if (value) return value;
-  }
-
-  const tableImages = findTomlSections(code, "images", false)[0];
-  if (tableImages) {
-    const binding = findTomlAssignmentInRange(
-      code,
-      "binding",
-      tableImages.bodyStart,
-      tableImages.bodyEnd,
-    );
-    const value = binding ? parseTomlString(binding.value) : undefined;
-    if (value) return value;
-  }
-
-  return "IMAGES";
-}
-
-function ensureTomlImagesBinding(code: string): string {
-  const inlineImages = findTopLevelTomlAssignment(code, "images");
-  const tableImages = findTomlSections(code, "images", false)[0];
-  if (inlineImages && tableImages) throw new Error("Wrangler TOML defines images twice.");
-
-  if (inlineImages) {
-    if (!isInlineTomlObject(inlineImages.value)) {
-      throw new Error("Expected top-level images to be a TOML object.");
-    }
-    const binding = findInlineTomlProperty(inlineImages.value, "binding");
-    const value = binding ? parseTomlString(binding.value) : undefined;
-    if (value) return code;
-    return setInlineTomlPropertyValue(code, inlineImages, "binding", '"IMAGES"');
-  }
-
-  if (tableImages) {
-    const binding = findTomlAssignmentInRange(
-      code,
-      "binding",
-      tableImages.bodyStart,
-      tableImages.bodyEnd,
-    );
-    const value = binding ? parseTomlString(binding.value) : undefined;
-    if (value) return code;
-    return setTomlTableAssignmentValue(code, tableImages, "binding", '"IMAGES"');
-  }
-
-  return appendTomlTopLevelAssignment(code, 'images = { binding = "IMAGES" }');
-}
-
-function findTomlKvNamespaceSection(code: string): TomlSection | undefined {
-  return findTomlSections(code, "kv_namespaces", true).find((section) => {
-    const binding = findTomlAssignmentInRange(code, "binding", section.bodyStart, section.bodyEnd);
-    return binding ? parseTomlString(binding.value) === "VINEXT_KV_CACHE" : false;
-  });
-}
-
-function ensureTomlKvNamespace(code: string): string {
-  if (findTomlKvNamespaceSection(code)) return code;
-  return appendTomlArrayTable(
-    code,
-    '[[kv_namespaces]]\nbinding = "VINEXT_KV_CACHE"\nid = "<your-kv-namespace-id>"',
-  );
-}
-
-function tomlKvNamespaceNeedsId(code: string): boolean {
-  const section = findTomlKvNamespaceSection(code);
-  if (!section) return true;
-  const id = findTomlAssignmentInRange(code, "id", section.bodyStart, section.bodyEnd);
-  const value = id ? parseTomlString(id.value) : undefined;
-  return !value || value === "<your-kv-namespace-id>";
-}
-
-function updateWranglerTomlConfigForCloudflare(
-  code: string,
-  options: CloudflareInitOptions,
-): string {
-  let output = code;
-  if (options.cdnCache === "workers-cache") {
-    output = ensureTomlCacheEnabled(output);
-  }
-  if (options.imageOptimization === "cloudflare-images") {
-    output = ensureTomlImagesBinding(output);
-  }
-  if (options.dataCache === "kv") {
-    output = ensureTomlKvNamespace(output);
-  }
-  return output;
 }
 
 function cacheImports(options: CloudflareInitOptions): string[] {
